pyproject.toml
[project]
name = "dog-pose-mvp"
version = "0.1.0"
description = "Ultralytics dog pose training script and Streamlit MVP."
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "numpy>=2.2.0",
    "opencv-python-headless>=4.11.0.86",
    "pandas>=2.2.0",
    "pillow>=11.1.0",
    "streamlit>=1.43.0",
    "ultralytics>=8.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dog_pose_mvp"]
